Enchance UI + Enhance Security + Refactor code
Next update will focus on google drive
so instead of upload to PostgreSQL
Will upload to google drive
the postgreSQL will only store the link

+++ REST-Conference/src/main/java/com/myapp/restapi/researchconference/entity/Review/Review.java
@@ -21,7 +21,7 @@ public class Review {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "reviewID")
-    private long reviewID;
+    private int reviewID;
 
     @OneToOne
     @JoinColumn(name = "bidID")
